added checks for the correct virtualenv (was bitten) and cleaned up reset script.

+++ dev/reset-trough-dev.sh
@@ -1,9 +1,17 @@
 #!/bin/bash
 
-venv=/Users/nlevitt/workspace/trough/trough-ve3
-source $venv/bin/activate
+if [ -z "$VIRTUAL_ENV" ] ; then
+    echo '$VIRTUAL_ENV is not set (please activate your trough virtualenv)'
+    exit 1
+fi
 
-script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+python -c 'import trough'
+if [ $? -ne 0 ]; then
+    echo "trough module could not be imported. Are you in the right virtualenv?"
+    exit 1
+fi
+
+script_dir=$(dirname $VIRTUAL_ENV)/dev
 $script_dir/stop-trough-dev.sh
 
 rm -vrf /tmp/trough-*.out
